pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaos-cal"
version = "2.0.0"
description = "CHAOS - Chain-based Algebraic Optimal Solver for Radio Interferometry Calibration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Arpan Pal"}
]
keywords = [
    "radio astronomy",
    "interferometry", 
    "calibration",
    "Jones matrices",
    "RIME",
    "CASA"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21",
    "scipy>=1.7",
    "python-casacore>=3.5",
    "h5py>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = ["cupy>=11.0"]
dev = ["pytest>=7.0", "pytest-cov", "black", "ruff"]

[project.scripts]
chaos = "chaos.cli.main:main"

[project.urls]
Homepage = "https://github.com/arpanpal/chaos"
Repository = "https://github.com/arpanpal/chaos"

[tool.setuptools.packages.find]
include = ["chaos*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["chaos/tests"]
python_files = ["test_*.py"]
